refactor: 기존 SortOption 이름 변경
