+++ src/main/java/be/dash/dashserver/core/domain/lesson/service/LessonService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import be.dash.dashserver.core.domain.common.Genre;
 import be.dash.dashserver.core.domain.common.Level;
-import be.dash.dashserver.core.domain.common.SortOption;
+import be.dash.dashserver.core.domain.lesson.LessonSortOption;
 import be.dash.dashserver.core.domain.lesson.Lessons;
 import lombok.RequiredArgsConstructor;
 
@@ -16,7 +16,7 @@ public class LessonService {
 
     private final LessonRepository lessonRepository;
 
-    public Lessons search(Genre genre, Level level, LocalDateTime startDateTime, LocalDateTime endDateTime, SortOption sortOption) {
+    public Lessons search(Genre genre, Level level, LocalDateTime startDateTime, LocalDateTime endDateTime, LessonSortOption sortOption) {
         Lessons lessons = new Lessons(
                 lessonRepository.findActiveLessonsByFilters(genre, level, startDateTime, endDateTime, LocalDateTime.now())
         );
